LPS-60022 move displayContext logic to the new displayContext
