+++ modules/apps/product-navigation/product-navigation-product-menu-web/src/main/java/com/liferay/product/navigation/product/menu/web/display/context/ProductMenuDisplayContext.java
@@ -21,22 +21,16 @@ import com.liferay.application.list.constants.ApplicationListWebKeys;
 import com.liferay.application.list.constants.PanelCategoryKeys;
 import com.liferay.application.list.display.context.logic.PanelCategoryHelper;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.model.Group;
 import com.liferay.portal.model.Layout;
-import com.liferay.portal.model.User;
 import com.liferay.portal.theme.ThemeDisplay;
-import com.liferay.portal.util.PortalUtil;
 
 import java.util.List;
 
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
-import javax.portlet.PortletURL;
 
 /**
  * @author Julio Camarero
@@ -59,10 +53,6 @@ public class ProductMenuDisplayContext {
 		_panelCategoryRegistry =
 			(PanelCategoryRegistry)_portletRequest.getAttribute(
 				ApplicationListWebKeys.PANEL_CATEGORY_REGISTRY);
-
-		_panelCategoryHelper =
-			(PanelCategoryHelper)_portletRequest.getAttribute(
-				ApplicationListWebKeys.PANEL_CATEGORY_HELPER);
 	}
 
 	public List<PanelCategory> getChildPanelCategories() {
@@ -75,33 +65,6 @@ public class ProductMenuDisplayContext {
 			_themeDisplay.getScopeGroup());
 
 		return _childPanelCategories;
-	}
-
-	public Group getMySiteGroup() throws PortalException {
-		if (_mySiteGroup != null) {
-			return _mySiteGroup;
-		}
-
-		User user = _themeDisplay.getUser();
-
-		List<Group> mySiteGroups = user.getMySiteGroups(
-			new String[] {User.class.getName()}, 1);
-
-		if (mySiteGroups.isEmpty()) {
-			return null;
-		}
-
-		_mySiteGroup = mySiteGroups.get(0);
-
-		return _mySiteGroup;
-	}
-
-	public String getMySiteGroupURL(boolean privateLayout)
-		throws PortalException {
-
-		Group mySiteGroup = getMySiteGroup();
-
-		return getMySiteGroupURL(mySiteGroup, privateLayout);
 	}
 
 	public String getRootPanelCategoryKey() {
@@ -141,19 +104,6 @@ public class ProductMenuDisplayContext {
 		return _rootPanelCategoryKey;
 	}
 
-	public boolean showMySiteGroup(boolean privateLayout)
-		throws PortalException {
-
-		Group mySiteGroup = getMySiteGroup();
-
-		if (mySiteGroup == null) {
-			return false;
-		}
-
-		return mySiteGroup.isShowSite(
-			_themeDisplay.getPermissionChecker(), privateLayout);
-	}
-
 	public boolean showProductMenu() {
 		Layout layout = _themeDisplay.getLayout();
 
@@ -173,44 +123,8 @@ public class ProductMenuDisplayContext {
 		return true;
 	}
 
-	protected String getGroupAdministrationURL(Group group) {
-		PortletURL groupAdministrationURL = null;
-
-		String portletId = _panelCategoryHelper.getFirstPortletId(
-			PanelCategoryKeys.SITE_ADMINISTRATION,
-			_themeDisplay.getPermissionChecker(), group);
-
-		if (Validator.isNotNull(portletId)) {
-			groupAdministrationURL = PortalUtil.getControlPanelPortletURL(
-				_portletRequest, group, portletId, 0, 0,
-				PortletRequest.RENDER_PHASE);
-
-			if (groupAdministrationURL != null) {
-				return groupAdministrationURL.toString();
-			}
-		}
-
-		return null;
-	}
-
-	protected String getMySiteGroupURL(Group group, boolean privateLayout) {
-		String groupDisplayURL = group.getDisplayURL(
-			_themeDisplay, privateLayout);
-
-		if (Validator.isNotNull(groupDisplayURL)) {
-			return groupDisplayURL;
-		}
-
-		return getGroupAdministrationURL(group);
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		ProductMenuDisplayContext.class);
-
 	private List<PanelCategory> _childPanelCategories;
-	private Group _mySiteGroup;
 	private final PanelAppRegistry _panelAppRegistry;
-	private final PanelCategoryHelper _panelCategoryHelper;
 	private final PanelCategoryRegistry _panelCategoryRegistry;
 	private final PortletRequest _portletRequest;
 	private final PortletResponse _portletResponse;
